fail if no host playbook
This is a useful playbook if you want to run a same playbook in a Tower workflow with different inventory and want to fail if the host is not present on the inventory.

There is an option in ansible Tower starting 3.4 that will allow to force the inventory at the workflow level. this will help avoiding the use of this trick but it can still be useful for other needs.